backend/requirements.txt
attrs==21.2.0
backports.entry-points-selectable==1.1.0
click==8.0.1
colorama==0.4.4
distlib==0.3.3
filelock==3.3.0
Flask==2.0.1
Flask-Cors==3.0.10
Flask-SQLAlchemy==2.5.1
greenlet==1.1.1
gunicorn==20.1.0
importlib-metadata==4.8.1
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
numpy==1.21.2
packaging==21.0
pandas==1.3.3
platformdirs==2.4.0
pluggy==1.0.0
py==1.10.0
pybind11==2.7.1
pyparsing==2.4.7
python-dateutil==2.8.2
python-dotenv==0.19.1
pytz==2021.1
scipy==1.7.1
six==1.16.0
SQLAlchemy==1.4.24
scikit-learn==1.0.1
toml==0.10.2
tox==3.24.4
typing-extensions==3.10.0.2
virtualenv==20.8.1
Werkzeug==2.0.1
psycopg2==2.7.7
flasgger==0.9.5
requests==2.24.0
retry==0.9.2
faiss-cpu==1.7.2
pytest==6.2.5
pytest-flask==1.2.0
